Removes Koa (and related deps) as a server option
