+++ README.md
@@ -95,11 +95,11 @@ The application structure presented in this boilerplate is **fractal**, where fu
 ├── blueprints               # Blueprint files for redux-cli
 ├── build                    # All build-related configuration
 │   └── webpack              # Environment-specific configuration files for webpack
-├── client                   # Koa application (uses webpack middleware)
-│   └── main.js              # Client entry and rendering
+├── client                   # Contains HTML templates used during build compilation
+│   └── index.js             # Client entry and rendering
 ├── config                   # Project configuration settings
-├── server                   # Koa application (uses webpack middleware)
-│   └── main.js              # Server application entry point
+├── server                   # Express application (uses webpack middleware)
+│   └── index.js             # Server application entry point
 ├── universal                # Shared application source code
 │   ├── components           # Reusable Presentational Components
 │   ├── containers           # Reusable Container Components
@@ -150,7 +150,7 @@ We're currently evaluating [Ava](https://github.com/avajs/ava) - which is instal
 Out of the box, this starter kit is deployable by serving the `~/dist` folder generated by `npm run deploy` (make sure to specify your target `NODE_ENV` as well).
 
 ### Static Deployments
-If you are serving the application via a web server such as nginx, make sure to direct incoming routes to the root `~/dist/index.html` file and let react-router take care of the rest. We are planning on enhancing the Koa server that comes with the starter kit to be extended to serve as an API and server rendered React components, so if you're interested in helping - feel free to jump in!
+If you are serving the application via a web server such as nginx, make sure to direct incoming routes to the root `~/dist/index.html` file and let react-router take care of the rest. We are planning on enhancing the Express server that comes with the starter kit to be extended to serve as an API and server rendered React components, so if you're interested in helping - feel free to jump in!
 
 #### Publish With Surge
 If you want to simply publish the site as a static client, the project is configured to use [surge.sh](http://surge.sh/).
@@ -282,7 +282,7 @@ You can browser additional plugins at [postcss.parts](http://postcss.parts/)
 
 ### Server
 
-This booster kit comes packaged with an Koa server. It's important to note that the sole purpose of this server is to provide `webpack-dev-middleware` and `webpack-hot-middleware` for hot module replacement. Using a custom Koa app in place of [webpack-dev-server](https://github.com/webpack/webpack-dev-server) makes it easier to extend the starter kit to include functionality such as API's, universal rendering, and more -- all without bloating the base boilerplate.
+This booster kit comes packaged with an Express server. It's important to note that the sole purpose of this server is to provide `webpack-dev-middleware` and `webpack-hot-middleware` for hot module replacement. Using a custom Express app in place of [webpack-dev-server](https://github.com/webpack/webpack-dev-server) makes it easier to extend the starter kit to include functionality such as API's, universal rendering, and more -- all without bloating the base boilerplate.
 
 ### Production Optimization
