+++ bin/express.js
@@ -1,8 +1,8 @@
 import config from '../config'
-import server from '../server/express'
+import server from '../server'
 import _debug from 'debug'
 
-const debug = _debug('app:bin:server:express')
+const debug = _debug('app:bin:server')
 const port = config.server_port
 const host = config.server_host
