+++ server/index.js
@@ -1,30 +1,17 @@
-import Koa from 'koa'
-import convert from 'koa-convert'
-import webpack from 'webpack'
-import webpackConfig from '../../build/webpack.config'
-import historyApiFallback from 'koa-connect-history-api-fallback'
-import serve from 'koa-static'
-import proxy from 'koa-proxy'
+import express from 'express'
 import _debug from 'debug'
-import config from '../../config'
-import webpackDevMiddleware from '../middleware/webpack-dev'
-import webpackHMRMiddleware from '../middleware/webpack-hmr'
+import webpack from 'webpack'
+import webpackConfig from '../build/webpack.config'
+import config from '../config'
 
-const debug = _debug('app:server:koa')
+const debug = _debug('app:server:express')
+const app = express()
 const paths = config.utils_paths
-const app = new Koa()
-
-// Enable koa-proxy if it has been enabled in the config.
-if (config.proxy && config.proxy.enabled) {
-  app.use(convert(proxy(config.proxy.options)))
-}
 
 // This rewrites all routes requests to the root /index.html file
-// (ignoring file requests). If you want to implement isomorphic
+// (ignoring file requests). If you want to implement universal
 // rendering, you'll want to remove this middleware.
-app.use(convert(historyApiFallback({
-  verbose: false
-})))
+app.use(require('connect-history-api-fallback')())
 
 // ------------------------------------
 // Apply Webpack HMR Middleware
@@ -32,17 +19,23 @@ app.use(convert(historyApiFallback({
 if (config.env === 'development') {
   const compiler = webpack(webpackConfig)
 
-  // Enable webpack-dev and webpack-hot middleware
-  const { publicPath } = webpackConfig.output
-
-  app.use(webpackDevMiddleware(compiler, publicPath))
-  app.use(webpackHMRMiddleware(compiler))
+  debug('Enable webpack dev and HMR middleware')
+  app.use(require('webpack-dev-middleware')(compiler, {
+    publicPath: webpackConfig.output.publicPath,
+    contentBase: paths.client(),
+    hot: true,
+    quiet: config.compiler_quiet,
+    noInfo: config.compiler_quiet,
+    lazy: false,
+    stats: config.compiler_stats
+  }))
+  app.use(require('webpack-hot-middleware')(compiler))
 
   // Serve static assets from ~/src/static since Webpack is unaware of
   // these files. This middleware doesn't need to be enabled outside
   // of development since this directory will be copied into ~/dist
   // when the application is compiled.
-  app.use(serve(paths.universal('static')))
+  app.use(express.static(paths.universal('static')))
 } else {
   debug(
     'Server is being run outside of live development mode, meaning it will ' +
@@ -55,7 +48,7 @@ if (config.env === 'development') {
   // Serving ~/dist by default. Ideally these files should be served by
   // the web server and not the app server, but this helps to demo the
   // server in production.
-  app.use(serve(paths.dist()))
+  app.use(express.static(paths.dist()))
 }
 
 export default app
